Add common module.
