+++ application-common/src/main/java/com/seo/dto/error/ErrorResponseDto.java
