+++ application-common/src/main/java/com/seo/dto/report/BrokenPagesReportResponse.java
