+++ application-crawler/src/main/java/com/seo/crawl/CrawlStatus.java
@@ -1,4 +1,4 @@
-package com.seo.dto.crawl;
+package com.seo.crawl;
 
 import com.seo.crawler.impl.CrawlTask;
 import lombok.AllArgsConstructor;
@@ -10,9 +10,11 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Data
 @Builder
-public class BasicCrawlStatus {
-    private String taskId;
-    private String domain;
+public class CrawlStatus {
     private String projectName;
+    private String domain;
+    private long crawledPages;
+    private long remainedPages;
+    private long brokenPagesCount;
     private CrawlTask.Status status;
 }
