+++ pom.xml
@@ -19,6 +19,7 @@
         <module>application-api</module>
         <module>application-service</module>
         <module>application-config</module>
+        <module>application-common</module>
     </modules>
     <properties>
         <maven-compiler-plugin.version>3.10.1</maven-compiler-plugin.version>
